+++ jannovar-filter/src/main/java/de/charite/compbio/jannovar/JannovarFilterCommandLineParser.java
@@ -47,8 +47,8 @@ public class JannovarFilterCommandLineParser {
 			result.verbosity = 3;
 
 		String args[] = cmd.getArgs(); // get remaining arguments
-		if (args.length != 3)
-			throw new ParseException("must exactly three none-option arguments, had: " + args.length);
+		if (args.length != 4)
+			throw new ParseException("must have exactly four none-option arguments, had: " + args.length);
 
 		if (cmd.getOptionValue("inheritance-mode") != null)
 			result.modeOfInheritance = ModeOfInheritance.valueOf(ModeOfInheritance.class,
@@ -58,12 +58,10 @@ public class JannovarFilterCommandLineParser {
 		if (cmd.getOptionValue("database") != null)
 			result.jannovarDB = cmd.getOptionValue("database");
 
-		if (result.geneWise && result.jannovarDB == null)
-			throw new ParseException("Enabled -g/--gene-wise mode but provided no -d/--database option!");
-
-		result.pedPath = args[0];
-		result.inputPath = args[1];
-		result.outputPath = args[2];
+		result.jannovarDB = args[0];
+		result.pedPath = args[1];
+		result.inputPath = args[2];
+		result.outputPath = args[3];
 
 		return result;
 	}
@@ -77,7 +75,6 @@ public class JannovarFilterCommandLineParser {
 		options.addOption(new Option("m", "inheritance-mode", true, "enable very verbose output"));
 		options.addOption(new Option("g", "gene-wise", false,
 				"gene-wise instead of variant-wise processing (required for compound heterozygous filtration)"));
-		options.addOption(new Option("d", "database", true, "path to Jannovar DB .ser file"));
 
 		parser = new GnuParser();
 	}
@@ -85,10 +82,10 @@ public class JannovarFilterCommandLineParser {
 	public void printHelp() {
 		final String HEADER = new StringBuilder().append("Jannovar Filter Tool")
 				.append("Use this command to filter VCF files.\n\n")
-				.append("Usage: java -jar jannovar-filter.jar [OPTIONS] <PED.ped> <IN.vcf> <OUT.vcf>\n\n").toString();
+				.append("Usage: java -jar jannovar-filter.jar <DB.ser> <PED.ped> <IN.vcf> <OUT.vcf>\n\n").toString();
 		final String FOOTER = new StringBuilder()
 		.append("\n\nExample: java -jar jannovar-filter.jar -m AUTOSOMAL_DOMINANT data/hg19_ucsc.ser fam.ped 123.vcf 123.filtered.vcf\n")
-				.append("         java -jar jannovar-filter.jar -g -d data/hg19_ucsc.ser -m AUTOSOMAL_RECESSIVE <PED.ped> <IN.vcf> <OUT.vcf>\n\n")
+				.append("         java -jar jannovar-filter.jar -g -m AUTOSOMAL_RECESSIVE data/hg19_ucsc.ser fam.ped 123.vcf 123.filtered.vcf\n\n")
 		.append("Diseases\n\n")
 		.append("The --inheritance-mode parameter can take one of the following values. When given")
 		.append("then the variants will be filtered to those being compatible with the given mode")